modesetting: Fix cursor disappeared when coordinates were negative
When the cursor is on the far left and top of the screen, cursor is
disappeared, and kernel prompt an error:

[drm:vop2_plane_atomic_check] *ERROR* Esmart2-win0 is invisible
(src: pos[0, 0] rect[64 x 64] dst: pos[1118466, 326] rect[64 x 64]

+++ hw/xfree86/drivers/modesetting/drmmode_display.c
@@ -1869,8 +1869,8 @@ drmmode_set_cursor_position(xf86CrtcPtr crtc, int x, int y)
 
     drmmode_crtc_bounds(crtc, &dx, &dy, &dw, &dh);
 
-    x = x * dw / crtc->mode.HDisplay + dx;
-    y = y * dh / crtc->mode.VDisplay + dy;
+    x = x * (int)dw / crtc->mode.HDisplay + dx;
+    y = y * (int)dh / crtc->mode.VDisplay + dy;
 
     drmModeMoveCursor(drmmode->fd, drmmode_crtc->mode_crtc->crtc_id, x, y);
 }
